Updated subtraction file listing to be more modular
