+++ client/src/js/subtraction/components/Detail/__tests__/Detail.test.js
@@ -21,7 +21,8 @@ describe("<SubtractionDetail />", () => {
             canModify: true,
             error: "",
             onGet: jest.fn(),
-            onShowRemove: jest.fn()
+            onShowRemove: jest.fn(),
+            files: []
         };
     });
 
@@ -60,4 +61,18 @@ describe("<SubtractionDetail />", () => {
         const wrapper = shallow(<SubtractionDetail {...props} />);
         expect(wrapper).toMatchSnapshot();
     });
+
+    // it("should render files when files is defined"),
+    //     () => {
+    //         props.files = [
+    //             {
+    //                 download_url: "/api/subtractions/xl8faqqz/files/subtraction.fa.gz",
+    //                 id: 22,
+    //                 name: "subtraction.fa.gz",
+    //                 size: 36461731,
+    //                 subtraction: "xl8faqqz",
+    //                 type: "fasta"
+    //             }
+    //         ];
+    //     };
 });
